+++ src/main/java/sk/stuba/fei/uim/oop/assignment3/shoppingCart/CartItem.java
@@ -6,8 +6,6 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
 import javax.persistence.Id;
 
 @Getter
@@ -18,5 +16,5 @@ import javax.persistence.Id;
 public class CartItem {
     @Id
     Long productId;
-    Long amount;
+    int amount;
 }
